control_citas/requirements.txt
Django==2.2
djangorestframework
markdown
python-dotenv
psycopg2-binary
